+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [v0.0.3-alpha]
 ### Added
+- 2019-09-18 : Allow load an entire dir of config files
 - 2019-09-16 : Support to custom workdir from deployment file
+- 2019-09-14 : Support to custom env per deployment
 - 2019-09-14 : Restart command
 ### Changed
 - 2019-09-16 : Now the system envs are loaded with command envs